flask/requirements.txt
flask
pandas
kafka-python
pyarrow
confluent-kafka[avro]
